pyproject.toml
[project]
name = "openttd-llm"
version = "0.1.0"
description = "OpenTTD AI with LLM integration - Generate Squirrel AI code using language models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "openai>=1.0",
    "anthropic>=0.18",
    "ollama>=0.1",
    "terminology>=1.0.16",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
]

# Note: Run with `uv run python llm_agent.py` or `uv run python openttd_admin.py`
# Entry points require package structure - these are provided for reference
# [project.scripts]
# openttd-admin = "openttd_admin:main"
# openttd-llm = "llm_agent:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]
